+++ Libraries/LibIPC/Decoder.cpp
@@ -178,7 +178,7 @@ ErrorOr<Core::ProxyData> decode(Decoder& decoder)
 {
     auto type = TRY(decoder.decode<Core::ProxyData::Type>());
     auto host_ipv4 = IPv4Address(TRY(decoder.decode<u32>()));
-    auto port = TRY(decoder.decode<int>());
+    auto port = TRY(decoder.decode<u16>());
 
     return Core::ProxyData { type, host_ipv4, port };
 }
